+++ CHANGELOG.md
@@ -1,15 +1,21 @@
-## [0.4.2] - 2018-10-31
+## [0.5.0] - 2018-11-01
+
+### Added
+- more field types: textarea, number, password, file
+- FieldTypeExample
+
+## [0.4.2] - 2018-10-30
 
 ### Added
 - storybook
 - dynamic form example
-- more field types: Radios, Checkboxes
+- more field types: radios, checkboxes
 - updated example
 
 ## [0.2.5] - 2018-10-28
 
 ### Added
-- Checkbox, Radio, Select types.
+- checkbox, radio, select types
 
 ## [0.1.3] - 2018-10-20